move fasthttp out of core library, and into integration package
