+++ v3/newrelic/segments.go
@@ -5,9 +5,6 @@ package newrelic
 
 import (
 	"net/http"
-
-	"github.com/valyala/fasthttp"
-	"github.com/valyala/fasthttp/fasthttpadaptor"
 )
 
 // SegmentStartTime is created by Transaction.StartSegmentNow and marks the
@@ -289,6 +286,23 @@ func (s *ExternalSegment) outboundHeaders() http.Header {
 	return outboundHeaders(s)
 }
 
+func (s *ExternalSegment) GetOutboundHeaders() http.Header {
+	return s.outboundHeaders()
+}
+
+// SetSecureAgentEvent allows integration packages to set the secureAgentEvent
+// for this external segment. That field is otherwise unexported and not available
+// for other manipulation.
+func (s *ExternalSegment) SetSecureAgentEvent(event any) {
+	s.secureAgentEvent = event
+}
+
+// GetSecureAgentEvent retrieves the secureAgentEvent previously stored by
+// a SetSecureAgentEvent method.
+func (s *ExternalSegment) GetSecureAgentEvent() any {
+	return s.secureAgentEvent
+}
+
 // StartSegmentNow starts timing a segment.
 //
 // Deprecated: StartSegmentNow is deprecated and will be removed in a future
@@ -332,36 +346,6 @@ func StartExternalSegment(txn *Transaction, request *http.Request) *ExternalSegm
 				request.Header.Set(key, value)
 			}
 		}
-		if IsSecurityAgentPresent() {
-			secureAgent.DistributedTraceHeaders(request, s.secureAgentEvent)
-		}
-	}
-
-	return s
-}
-
-func StartExternalSegmentFastHTTP(txn *Transaction, ctx *fasthttp.RequestCtx) *ExternalSegment {
-	if nil == txn {
-		txn = transactionFromRequestContextFastHTTP(ctx)
-	}
-	request := &http.Request{}
-
-	fasthttpadaptor.ConvertRequest(ctx, request, true)
-	s := &ExternalSegment{
-		StartTime: txn.StartSegmentNow(),
-		Request:   request,
-	}
-	if IsSecurityAgentPresent() {
-		s.secureAgentEvent = secureAgent.SendEvent("OUTBOUND", request)
-	}
-
-	if request != nil && request.Header != nil {
-		for key, values := range s.outboundHeaders() {
-			for _, value := range values {
-				request.Header.Set(key, value)
-			}
-		}
-
 		if IsSecurityAgentPresent() {
 			secureAgent.DistributedTraceHeaders(request, s.secureAgentEvent)
 		}
